backend/requirements.txt
asgiref==3.8.1
Django==3.2.23
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
neo4j==5.12.0
PyJWT==2.10.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
